requirements.txt
sounddevice
numpy
openwakeword
requests
PyYAML
soundfile
piper
librosa